+++ packages/template-retail-react-app/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## v3.1.0-dev (Jun 25, 2024)
+### Bug Fixes
+
+- Out of stock and low stock items are removed from cart and checkout as unavailable products [#1865](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1865)
 
 ## v3.0.0 (Jun 25, 2024)